Fix volume and power button lock on Power+VolUp
There is a bug on fw 1.7 where the removed function enable_joypad() is now locking the power and volume buttons until a console restart. 
This is an unexpected behaviour, since it was working on fw 1.5.
This change will avoid to try to lock the joypad input and it will just disable the screen at the minimum brightness.